refactor looking good
