+++ packages/design/src/scripts/tokenServices/tokenConstants.ts
@@ -1,0 +1,1 @@
+export const baseUnit = "16px";
